+++ NEWS.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+## Changes:
+- Add `assay` parameter to `subset.Seurat` to specify the default assay for expression filters ([#292](https://github.com/satijalab/seurat-object/pull/292))
+
 # SeuratObject 5.4.0
 
 ## Changes: